browser: create fe api
