+++ forward_engineering/api/getDatabases.js
@@ -1,0 +1,10 @@
+const reApi = require('../../reverse_engineering/api');
+
+function getDatabases(connectionInfo, logger, callback, app) {
+	logger.progress({ message: 'Find all schemas' });
+	reApi.getSchemaNames(connectionInfo, logger, callback, app);
+}
+
+module.exports = {
+	getDatabases,
+};
